64 bit fixes (initial patch by Gwenole Beauchesne)
git-svn-id: svn://svn.savannah.nongnu.org/qemu/trunk@1458 c046a42c-6fe2-441c-8c8c-71466251a162

+++ ip_icmp.h
@@ -83,7 +83,7 @@ struct icmp {
             struct ip idi_ip;
             /* options and then 64 bits of data */
         } id_ip;
-        u_long id_mask;
+        uint32_t id_mask;
         char id_data[1];
     } icmp_dun;
 #define icmp_otime icmp_dun.id_ts.its_otime
